WIP: move batching to renderer
